requirements.txt
Django>=3.2
psycopg2-binary>=2.8,<3.0
celery==5.3.6
redis==5.0.1
djangorestframework
requests
djangorestframework-simplejwt